pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "client-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.34",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "cryptg>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]
